[chore] more bug fixes and refactorings
Former-commit-id: 130a5723f5069cc3e37f2b221c72d93c868075a6

+++ Graphics/IRenderer.h
@@ -102,19 +102,19 @@ namespace Graphics
 	inline Float 
 		IRenderer::GetWidth()
 	{
-		return mWidth;
+		return static_cast<Float>(mWidth);
 	}
 
 	inline Float 
 		IRenderer::GetHeight()
 	{
-		return mHeight;
+		return static_cast<Float>(mHeight);
 	}
 
 	inline Float 
 		IRenderer::GetAspect()
 	{
-		return mWidth / mHeight;
+		return static_cast<Float>(mWidth) / static_cast<Float>(mHeight);
 	}
 
 	inline Float
